+++ src/Utils/TSP/TemplateTSP.java
@@ -6,7 +6,7 @@ import java.util.Iterator;
 public abstract class TemplateTSP implements TSP {
     private Integer[] bestSol;
     protected TSPGraph g;
-    private int bestSolCost;
+    private float bestSolCost;
     private int timeLimit;
     private long startTime;
 
@@ -20,7 +20,7 @@ public abstract class TemplateTSP implements TSP {
         for (int i=1; i<g.getNbVertices(); i++) unvisited.add(i);
         Collection<Integer> visited = new ArrayList<Integer>(g.getNbVertices());
         visited.add(0); // The first visited vertex is 0
-        bestSolCost = Integer.MAX_VALUE;
+        bestSolCost = Float.MAX_VALUE;
         branchAndBound(0, unvisited, visited, 0);
     }
 
@@ -30,7 +30,7 @@ public abstract class TemplateTSP implements TSP {
         return -1;
     }
 
-    public int getSolutionCost(){
+    public float getSolutionCost(){
         if (g != null)
             return bestSolCost;
         return -1;
@@ -62,7 +62,7 @@ public abstract class TemplateTSP implements TSP {
      * @param currentCost the cost of the path corresponding to <code>visited</code>
      */
     private void branchAndBound(int currentVertex, Collection<Integer> unvisited,
-                                Collection<Integer> visited, int currentCost){
+                                Collection<Integer> visited, float currentCost){
         if (System.currentTimeMillis() - startTime > timeLimit) return;
         if (unvisited.size() == 0){
             if (g.isArc(currentVertex,0)){
